#211: Move get_list(_field)_delimiter from BaseModel to INIBasedModel
